+++ test/unit/js/frontend-tabletop.mjs
@@ -23,7 +23,7 @@
 import { expect } from 'chai'
 
 import {
-  _test,
+  _test as testState,
   setTableNo,
   getSetup,
   FLAGS
@@ -33,25 +33,32 @@ import {
   TYPE_SQUARE,
   TYPE_HEX,
   TYPE_HEX2,
-  populatePiecesDefaults
+  getFeatures,
+  populatePiecesDefaults,
+  findPiece
 } from '../../../src/js/view/room/tabletop/tabledata.mjs'
 
 import {
   selectionAdd,
-  selectionClear
+  selectionClear,
+  selectionGetFeatures,
+  selectionGetPieces
 } from '../../../src/js/view/room/tabletop/selection.mjs'
 
 import {
-  moveSelected
+  _test,
+  pileSelected,
+  moveSelected,
+  deleteSelected
 } from '../../../src/js/view/room/tabletop/index.mjs'
 
 /**
  * Initialize table+room data for tests.
  */
 function setupTestData () {
-  _test.setRoom(JSON.parse(roomJSON))
+  testState.setRoom(JSON.parse(roomJSON))
   for (let i = 1; i <= 9; i++) {
-    _test.setTable(i, [])
+    testState.setTable(i, [])
   }
   setTableNo(1, false)
 }
@@ -63,12 +70,14 @@ describe('Frontend - tabletop.mjs', function () {
 
   it('moveSelected() square', function () {
     getSetup().type = TYPE_SQUARE
-    _test.setTable(1, populatePiecesDefaults([
+    testState.setTable(1, populatePiecesDefaults([
       { ...JSON.parse(pieceJSON), id: '1', x: 32 + 64 * 0, y: 32 + 64 * 0 },
       { ...JSON.parse(pieceJSON), id: '2', x: 32 + 64 * 1, y: 32 + 64 * 1, f: FLAGS.NO_MOVE },
       { ...JSON.parse(pieceJSON), id: '3', x: 32 + 64 * 2, y: 32 + 64 * 2 }
     ]))
     expect(moveSelected(1, 2, false)).to.be.eql([])
+    expect(selectionGetPieces().length).to.be.eql(0)
+    expect(selectionGetFeatures().move).to.be.eql(false)
 
     selectionAdd('3')
     const move1 = moveSelected(1, 2, false)
@@ -76,6 +85,8 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move1[0].id).to.be.eql('3')
     expect(move1[0].x).to.be.eql(32 + 64 * 2 + 1 * 64)
     expect(move1[0].y).to.be.eql(32 + 64 * 2 + 2 * 64)
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().move).to.be.eql(true)
     selectionClear()
 
     // dont move protected
@@ -88,8 +99,13 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move3[0].id).to.be.eql('1')
     expect(move3[0].x).to.be.eql(32 + 64 * 0 + 1 * 64)
     expect(move3[0].y).to.be.eql(32 + 64 * 0 + 2 * 64)
+    expect(selectionGetPieces().length).to.be.eql(2)
+    expect(selectionGetFeatures().move).to.be.eql(true)
+    selectionClear()
 
     // move out of table
+    selectionAdd('1')
+    selectionAdd('2')
     selectionAdd('3')
     const move4 = moveSelected(1, 2, false)
     expect(move4.length).to.be.eql(2)
@@ -103,16 +119,21 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move8.length).to.be.eql(0)
     const move9 = moveSelected(1, 2, false)
     expect(move9.length).to.be.eql(2)
+    expect(selectionGetPieces().length).to.be.eql(3)
+    expect(selectionGetFeatures().move).to.be.eql(true)
+    selectionClear()
   })
 
   it('moveSelected() hex', function () {
     getSetup().type = TYPE_HEX
-    _test.setTable(1, populatePiecesDefaults([
+    testState.setTable(1, populatePiecesDefaults([
       { ...JSON.parse(pieceJSON), id: '4', x: 55, y: 32 },
       { ...JSON.parse(pieceJSON), id: '5', x: 32 + 64 * 1, y: 32 + 64 * 1, f: FLAGS.NO_MOVE },
       { ...JSON.parse(pieceJSON), id: '6', x: 110, y: 128 }
     ]))
     expect(moveSelected(1, 2, false)).to.be.eql([])
+    expect(selectionGetPieces().length).to.be.eql(0)
+    expect(selectionGetFeatures().move).to.be.eql(false)
 
     selectionAdd('4')
     const move1 = moveSelected(1, 0, false)
@@ -125,6 +146,8 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move2[0].id).to.be.eql('4')
     expect(move2[0].x).to.be.eql(55)
     expect(move2[0].y).to.be.eql(32 + 64)
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().move).to.be.eql(true)
     selectionClear()
 
     selectionAdd('6')
@@ -138,6 +161,8 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move4[0].id).to.be.eql('6')
     expect(move4[0].x).to.be.eql(110)
     expect(move4[0].y).to.be.eql(128 - 64)
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().move).to.be.eql(true)
     selectionClear()
 
     selectionAdd('4')
@@ -159,18 +184,26 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move6[1].id).to.be.eql('4')
     expect(move6[1].x).to.be.eql(55)
     expect(move6[1].y).to.be.eql(32 + 64)
-
+    expect(selectionGetPieces().length).to.be.eql(3)
+    expect(selectionGetFeatures().move).to.be.eql(true)
     selectionClear()
   })
 
   it('moveSelected() hex2', function () {
     getSetup().type = TYPE_HEX2
-    _test.setTable(1, populatePiecesDefaults([
+    testState.setTable(1, populatePiecesDefaults([
       { ...JSON.parse(pieceJSON), id: '7', x: 32, y: 55 },
       { ...JSON.parse(pieceJSON), id: '8', x: 32 + 64 * 1, y: 32 + 64 * 1, f: FLAGS.NO_MOVE },
       { ...JSON.parse(pieceJSON), id: '9', x: 128, y: 110 }
     ]))
     expect(moveSelected(1, 2, false)).to.be.eql([])
+    expect(selectionGetPieces().length).to.be.eql(0)
+    expect(selectionGetFeatures().move).to.be.eql(false)
+
+    selectionAdd('8')
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().move).to.be.eql(false)
+    selectionClear()
 
     selectionAdd('7')
     const move1 = moveSelected(0, 1, false)
@@ -183,6 +216,8 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move2[0].id).to.be.eql('7')
     expect(move2[0].x).to.be.eql(32 + 64)
     expect(move2[0].y).to.be.eql(55)
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().move).to.be.eql(true)
     selectionClear()
 
     selectionAdd('9')
@@ -196,6 +231,8 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move4[0].id).to.be.eql('9')
     expect(move4[0].x).to.be.eql(128 - 64)
     expect(move4[0].y).to.be.eql(110)
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().move).to.be.eql(true)
     selectionClear()
 
     selectionAdd('7')
@@ -217,7 +254,121 @@ describe('Frontend - tabletop.mjs', function () {
     expect(move6[1].id).to.be.eql('7')
     expect(move6[1].x).to.be.eql(32 + 64)
     expect(move6[1].y).to.be.eql(55)
+    expect(selectionGetPieces().length).to.be.eql(3)
+    expect(selectionGetFeatures().move).to.be.eql(true)
+    selectionClear()
+  })
 
+  it('clonePieces()', function () {
+    testState.setTable(1, populatePiecesDefaults([
+      { ...JSON.parse(pieceJSON), id: 'A', x: 64, y: 64 },
+      { ...JSON.parse(pieceJSON), id: 'B', x: 32, y: 32, f: FLAGS.NO_CLONE },
+      { ...JSON.parse(pieceJSON), id: 'C', x: 256, y: 128 }
+    ]))
+
+    const a = findPiece('A')
+    const b = findPiece('B')
+    const c = findPiece('C')
+
+    expect(_test.clonePieces([], { x: 64, y: 64 }, false)).to.be.eql([])
+    expect(getFeatures([]).clone).to.be.eql(false)
+
+    const clone1 = _test.clonePieces([a], { x: 64 + 64 * 10, y: 64 + 64 * 8 }, false) // will snap!
+    expect(clone1.length).to.be.eql(1)
+    expect(clone1[0].id).to.be.eql('A')
+    expect(clone1[0].x).to.be.eql(64 + 64 * 10)
+    expect(clone1[0].y).to.be.eql(64 + 64 * 8)
+    expect(getFeatures([a]).clone).to.be.eql(true)
+
+    const clone2 = _test.clonePieces([a, b], { x: 64 + 64 * 10, y: 64 + 64 * 8 }, false) // will snap!
+    expect(clone2.length).to.be.eql(1)
+    expect(clone2[0].id).to.be.eql('A')
+    expect(clone2[0].x).to.be.eql(64 + 64 * 10)
+    expect(clone2[0].y).to.be.eql(64 + 64 * 8)
+    expect(getFeatures([a, b]).clone).to.be.eql(true)
+
+    const clone3 = _test.clonePieces([a, b, c], { x: 64 * 10, y: 64 * 8 }, false) // will snap!
+    expect(clone3.length).to.be.eql(2)
+    expect(clone3[0].id).to.be.eql('A')
+    expect(clone3[0].x).to.be.eql(64 * 10 - (256 - 64) / 2)
+    expect(clone3[0].y).to.be.eql(64 * 8 - (128 - 64) / 2)
+    expect(clone3[1].id).to.be.eql('C')
+    expect(clone3[1].x).to.be.eql(64 * 10 + (256 - 64) / 2)
+    expect(clone3[1].y).to.be.eql(64 * 8 + (128 - 64) / 2)
+    expect(getFeatures([a, b, c]).clone).to.be.eql(true)
+
+    selectionClear()
+  })
+
+  it('pileSelected()', function () {
+    testState.setTable(1, populatePiecesDefaults([
+      { ...JSON.parse(pieceJSON), id: 'P1', x: 64, y: 64 },
+      { ...JSON.parse(pieceJSON), id: 'P2', x: 32 + 64 * 1, y: 32 + 64 * 1, f: FLAGS.NO_MOVE },
+      { ...JSON.parse(pieceJSON), id: 'P3', x: 256, y: 128 }
+    ]))
+    expect(selectionGetPieces().length).to.be.eql(0)
+    expect(selectionGetFeatures().pile).to.be.eql(false)
+    expect(pileSelected(undefined, false)).to.be.eql([])
+
+    selectionAdd('P1')
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().pile).to.be.eql(false)
+    expect(pileSelected(undefined, false)).to.be.eql([])
+    selectionClear()
+
+    selectionAdd('P1')
+    selectionAdd('P2')
+    expect(selectionGetPieces().length).to.be.eql(2)
+    expect(selectionGetFeatures().pile).to.be.eql(false)
+    expect(pileSelected(undefined, false)).to.be.eql([])
+    selectionClear()
+
+    selectionAdd('P1')
+    selectionAdd('P2')
+    selectionAdd('P3')
+    expect(selectionGetPieces().length).to.be.eql(3)
+    expect(selectionGetFeatures().pile).to.be.eql(true)
+    const pile1 = pileSelected(undefined, false)
+    expect(pile1.length).to.be.eql(2)
+    expect(pile1[0].id).to.be.eql('P1')
+    expect(pile1[0].x).to.be.eql(160)
+    expect(pile1[0].y).to.be.eql(96)
+    expect(pile1[1].id).to.be.eql('P3')
+    expect(pile1[1].x).to.be.eql(160)
+    expect(pile1[1].y).to.be.eql(96)
+    selectionClear()
+  })
+
+  it('deleteSelected()', function () {
+    testState.setTable(1, populatePiecesDefaults([
+      { ...JSON.parse(pieceJSON), id: 'P1', x: 64, y: 64 },
+      { ...JSON.parse(pieceJSON), id: 'P2', x: 32 + 64 * 1, y: 32 + 64 * 1, f: FLAGS.NO_DELETE },
+      { ...JSON.parse(pieceJSON), id: 'P3', x: 256, y: 128 },
+      { ...JSON.parse(pieceJSON), id: 'P4', x: 1256, y: 1128 }
+    ]))
+    expect(selectionGetPieces().length).to.be.eql(0)
+    expect(selectionGetFeatures().delete).to.be.eql(false)
+    expect(deleteSelected(false)).to.be.eql([])
+
+    selectionAdd('P2')
+    expect(selectionGetPieces().length).to.be.eql(1)
+    expect(selectionGetFeatures().delete).to.be.eql(false)
+    expect(deleteSelected(false)).to.be.eql([])
+    selectionClear()
+
+    selectionAdd('P1')
+    selectionAdd('P2')
+    expect(selectionGetPieces().length).to.be.eql(2)
+    expect(selectionGetFeatures().delete).to.be.eql(true)
+    expect(deleteSelected(false)).to.be.eql(['P1'])
+    selectionClear()
+
+    selectionAdd('P2')
+    selectionAdd('P3')
+    selectionAdd('P4')
+    expect(selectionGetPieces().length).to.be.eql(3)
+    expect(selectionGetFeatures().delete).to.be.eql(true)
+    expect(deleteSelected(false)).to.be.eql(['P3', 'P4'])
     selectionClear()
   })
 })