+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Download current and previous releases from [https://github.com/ludus-leonis/Fre
 ### Other changes
 
 * added system-assets separator line in library window
+* exclude protected pieces from multi-select operation, instead of denying it to all
 * improved display of availables sides and sizes in the library